# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Shovel Heroes Public API
  version: 0.2.0
  description: |
    REST endpoints replacing previous Base44 SDK usage.
    Comprehensive API documentation including authentication, admin panel, and all CRUD operations.
    Last updated: 2025-10-03 - Added auth, admin, and missing CRUD endpoints.
...
servers:
  - url: https://your.api.server
    description: Production
  - url: http://localhost:8787
    description: Local Dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Timestamp:
      type: string
      format: date-time
      description: ISO 8601 時間戳記 (UTC)
      example: "2025-10-02T08:12:30Z"
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: 錯誤訊息
          example: 未授權的請求
        code:
          type: string
          description: 內部錯誤代碼 (可選)
          example: AUTH_REQUIRED
      required: [message]
    UserRole:
      type: string
      enum: [volunteer, victim, ngo_coordinator, regional_admin, data_analyst, super_admin]
      description: 使用者角色
    UserStatus:
      type: string
      enum: [active, suspended, pending_verification, inactive]
      description: 使用者狀態
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user_role:
          $ref: '#/components/schemas/UserRole'
        action:
          type: string
          description: 操作類型
          example: UPDATE_USER_STATUS
        resource_type:
          type: string
          description: 資源類型
          example: users
        resource_id:
          type: string
          nullable: true
          description: 資源 ID
        ip_address:
          type: string
          nullable: true
        user_agent:
          type: string
          nullable: true
        request_data:
          type: object
          nullable: true
        old_value:
          type: object
          nullable: true
        new_value:
          type: object
          nullable: true
        created_at:
          $ref: '#/components/schemas/Timestamp'
      required: [id, action, created_at]
...
  responses:
    Unauthorized:
      description: 未授權 (需要 Bearer Token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 找不到資源
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /admin/users:
    get:
      summary: List all users
      description: 取得所有使用者清單 (含篩選與分頁), 需要 coordinator 以上權限
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStatus'
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Users list with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        user_id:
                          type: string
                          format: uuid
                        role:
                          $ref: '#/components/schemas/UserRole'
                        email:
                          type: string
                        phone_number:
                          type: string
                        status:
                          $ref: '#/components/schemas/UserStatus'
                        created_at:
                          $ref: '#/components/schemas/Timestamp'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Insufficient permissions
  /admin/users/{user_id}:
    get:
      summary: Get user details
      description: 取得使用者詳細資訊
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete user (soft delete)
      description: 軟刪除使用者 (僅 super_admin)
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Requires super_admin role
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/users/{user_id}/status:
    patch:
      summary: Update user status
      description: 更新使用者狀態 (suspend/activate)
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/UserStatus'
                reason:
                  type: string
              required: [status]
      responses:
        '200':
          description: Status updated
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/verify-victim:
    post:
      summary: Verify victim identity
      description: 驗證災民身份與狀態
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  format: uuid
                action:
                  type: string
                  enum: [approve, reject, need_more_info]
                notes:
                  type: string
              required: [user_id, action]
      responses:
        '200':
          description: Verification processed
        '400':
          description: User is not a victim
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/audit-logs:
    get:
      summary: Get audit logs
      description: 取得稽核日誌 (含篩選)
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Audit logs with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  pagination:
                    type: object
  /admin/audit-logs/export:
    get:
      summary: Export audit logs as CSV
      description: 匯出稽核日誌為 CSV 檔案
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /admin/stats:
    get:
      summary: Get dashboard statistics
      description: 取得管理後台統計資料
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: object
                    properties:
                      total:
                        type: integer
                      byRole:
                        type: array
                        items:
                          type: object
                          properties:
                            role:
                              $ref: '#/components/schemas/UserRole'
                            count:
                              type: integer
                      byStatus:
                        type: array
                        items:
                          type: object
                          properties:
                            status:
                              $ref: '#/components/schemas/UserStatus'
                            count:
                              type: integer
                      activeVolunteers:
                        type: integer
                      pendingVerifications:
                        type: integer
                      recentSignups:
                        type: integer
  /admin/recent-activity:
    get:
      summary: Get recent activity feed
      description: 取得最近的活動記錄
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Recent activities
...